s64 <-> u32 swapped

+++ include/revolution/os.h
@@ -68,8 +68,8 @@ bool OSIsEnableInterrupts();
 /*
 * Real-time clock API
 */
-s64 OSGetTick();
-u32 OSGetTime();
+u32 OSGetTick();
+s64 OSGetTime();
 
 #ifdef __cplusplus
 }
